* Source/NSPortCoder.m (-versionForClassName:): Use NSInteger for   version to avoid compiler warning on 64-bit system.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@34308 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSPortCoder.m
@@ -1925,7 +1925,7 @@ static IMP	_xRefImp;	/* Serialize a crossref.	*/
 - (NSInteger) versionForClassName: (NSString*)className
 {
   GSClassInfo	*info = nil;
-  unsigned	version = NSNotFound;
+  NSInteger	version = NSNotFound;
   unsigned	count = GSIArrayCount(_clsAry);
 
   /*
